pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-session-harness"
version = "0.1.0"
description = "Portable hook/journal and binding-first worker launcher for Claude and Codex."
requires-python = ">=3.11"
dependencies = []

[project.optional-dependencies]
yaml = ["PyYAML>=6.0"]

[tool.pytest.ini_options]
testpaths = []
